added error code returns to all sasa functions
Had to change the type declarations for all the functions from void to
int.

Instead of crashing the command will return a -1. If it succeeds it
will return a 1.

When the python shrake_rupley command receives the error code it will
raise a ValueError if it receives a -1 and continue if it gets
anything else.